+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bootstrapper
 * Update to use PSR-4
 * Remove Form::append* etc in favour of an InputGroup class
 * Remove Form::control_group in favour of a ControlGroup class
+* Rename Progress to ProgressBar
 
 4.0.0
 =====